+++ packages/react-native-audio-api/common/cpp/audioapi/core/sources/AudioBufferQueueSourceNode.cpp
@@ -279,7 +279,7 @@ void AudioBufferQueueSourceNode::processWithInterpolation(
       }
     }
 
-    for (int i = 0; i < processingBuffer->getNumberOfChannels(); i += 1) {
+    for (size_t i = 0; i < processingBuffer->getNumberOfChannels(); i += 1) {
       const auto destination = processingBuffer->getChannel(i)->span();
       const auto currentSource = buffer->getChannel(i)->span();
